adjustmentes in bussiness page
